Attempt to modernize PlayNote... run into problems due to MKSynthPatches needing MKUnitGenerators, but MKUnitGenerators doesn't compile.

+++ MusicKit/Frameworks/MKSynthPatches/_MKList.h
@@ -25,7 +25,7 @@
     id *theList;
 }
 -(id)init;
--(id)objectAtIndex:(int)indx;
+-(id)objectAtIndex:(NSInteger)indx;
 -(id *)baseAddress;
 
 @end
